refator: render work item hook
