+++ package.json
@@ -16,6 +16,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-draggable": "^4.4.4",
+    "react-masonry-css": "^1.0.16",
     "react-moment": "^1.1.1",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
