+++ src/module/work-management/component/bucket-board/component/bucket-list/style.scss
@@ -37,7 +37,6 @@
             }
             .bucket-card {
                 width: 100%;
-                height: 100%;
                 padding: 10px;
                 box-shadow: 0px 0px 2px 0px grey;
                 background-color: white;
